create a UMD hybrid module using webpack
