+++ client/javascript/package.json
@@ -1,11 +1,11 @@
 {
   "name": "datajunction",
-  "version": "0.0.1",
+  "version": "0.0.2-rc.2",
   "description": "A Javascript client for interacting with a DataJunction server",
-  "main": "index.js",
+  "module": "src/index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "babel src -d dist",
+    "build": "rm -rf dist/* && webpack && babel src -d dist",
     "lint": "eslint src --ext .js",
     "lint:fix": "npm run lint -- --fix"
   },
@@ -33,6 +33,8 @@
     "eslint-config-standard": "^17.0.0",
     "eslint-plugin-import": "^2.27.5",
     "eslint-plugin-n": "^15.7.0",
-    "eslint-plugin-promise": "^6.1.1"
+    "eslint-plugin-promise": "^6.1.1",
+    "webpack": "^5.81.0",
+    "webpack-cli": "^5.0.2"
   }
 }
