+++ client/javascript/webpack.config.js
@@ -1,0 +1,23 @@
+const path = require('path');
+
+module.exports = {
+  entry: './src/index.js',
+  mode: "production",
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'datajunction.js',
+    globalObject: 'this',
+    library: {
+      name: 'datajunction',
+      type: 'umd',
+    },
+  },
+  externals: {
+    lodash: {
+      commonjs: 'lodash',
+      commonjs2: 'lodash',
+      amd: 'lodash',
+      root: '_',
+    },
+  },
+};
