+++ docs/content/0.1.0/docs/getting-started/clients.md
@@ -20,11 +20,31 @@ print(dj.metrics())
 
 ## Javascript
 
-The DataJunction javascript client can be installed using [npm](https://www.npmjs.com/).
+The DataJunction javascript client can be installed using [npm](https://www.npmjs.com/) for use in a node project
+or using the [UNPKG](https://www.unpkg.com/) CDN for client-side use.
+
+
+
+
+{{< tabs "javascript client" >}}
+{{< tab "CommonJS" >}}
 ```sh
 npm install datajunction
 ```
 ```js
-const dj = new DJClient("http://localhost:8000")
-dj.metrics.get().then(data => console.log(data))
+const datajunction = require('datajunction')
+
+const dj = new datajunction.DJClient('http://localhost:8000')
+dj.nodes.get().then(data => console.log(data))
 ```
+{{< /tab >}}
+{{< tab "ES6" >}}
+```html
+<script src="https://unpkg.com/datajunction/dist/datajunction.js"></script>
+<script>
+    const dj = new window.datajunction.DJClient("http://localhost:8000");
+    dj.nodes.get().then(data => console.log(data));
+</script>
+```
+{{< /tab >}}
+{{< /tabs >}}
